fix manage sessions auth
