+++ frontend/src/api.js
@@ -1,0 +1,26 @@
+import axios from "axios";
+export const API_URL = "http://localhost:3001";
+
+const api = axios.create({ baseURL: API_URL });
+
+api.interceptors.request.use(
+  config => {
+    const token = localStorage.getItem("token");
+    if (token) config.headers.Authorization = "Bearer " + token;
+    return config;
+  },
+  error => Promise.reject(error)
+);
+
+api.interceptors.response.use(
+  res => res,
+  err => {
+    if (err.response && err.response.status === 401) {
+      localStorage.removeItem("token");
+      window.location.reload();
+    }
+    return Promise.reject(err);
+  }
+);
+
+export default api;
